Colors now use 0.-1. instead of 0-0 or 0-255

+++ Packages/vcs/Lib/VTKPlots.py
@@ -687,7 +687,7 @@ class VTKVCSBackend(object):
             cmap = vcs.getcolormap(contLine.colormap)
         else:
             cmap = self.canvas.getcolormap()
-        color = [int(c / 100.) for c in cmap.index[contLine.color[0]]]
+        color = [c / 100. for c in cmap.index[contLine.color[0]]]
         line_prop.SetColor(*color)
 
         # Stippling
